requirements.txt
# Monday CRM Agent
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-telegram-bot>=20.8
google-generativeai>=0.7.0
httpx>=0.26.0
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
pytz>=2024.1
